component folders -re-organized component folders
