+++ src/components/Pages/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Footer from './Footer';
-import Nav from './Nav';
+import Footer from '../Footer/Footer';
+import Nav from '../NavBar/Nav';
 import PhotoLeft from '../../assets/painting.jpeg';
 import PhotoRight from '../../assets/workshop.jpeg';
